backend/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-filter==22.1
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
Pillow==9.4.0
python-dotenv==1.0.0
drf-extra-fields==3.4.1